+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Unreleased
+
+
+### Changed
+
+- Add "warn" issue level and update issue messages
+
+
+---
+
+
 ## v0.4.0 – 2025-05-18